+++ test/releasy/packagers_test.rb
@@ -3,11 +3,11 @@ require File.expand_path("../teststrap", File.dirname(__FILE__))
 # Test all packagers at once, since they are pretty much identical.
 [
     [:dmg,     Releasy::Packagers::Dmg,      %[GZIP=-9 hdiutil create -fs HFS+ -srcfolder "f" -volname "Test App 0.1" "f.dmg"]],
-    [:exe,     Releasy::Packagers::Exe,      %[7z a -mmt -bd -t7z -mx9 -sfx7z.sfx "f.exe" "f"]],
-    [:"7z",    Releasy::Packagers::SevenZip, %[7z a -mmt -bd -t7z -mx9 "f.7z" "f"]],
-    [:tar_bz2, Releasy::Packagers::TarBzip2, %[7z a -so -mmt -bd -ttar "f.tar" "f" | 7z a -si -bd -tbzip2 -mx9 "f.tar.bz2"]],
-    [:tar_gz,  Releasy::Packagers::TarGzip,  %[7z a -so -mmt -bd -ttar "f.tar" "f" | 7z a -si -bd -tgzip -mx9 "f.tar.gz"]],
-    [:zip,     Releasy::Packagers::Zip,      %[7z a -mmt -bd -tzip -mx9 "f.zip" "f"]],
+    [:exe,     Releasy::Packagers::Exe,      %[7za a -mmt -bd -t7z -mx9 -sfx7z.sfx "f.exe" "f"]],
+    [:"7z",    Releasy::Packagers::SevenZip, %[7za a -mmt -bd -t7z -mx9 "f.7z" "f"]],
+    [:tar_bz2, Releasy::Packagers::TarBzip2, %[7za a -so -mmt -bd -ttar "f.tar" "f" | 7za a -si -bd -tbzip2 -mx9 "f.tar.bz2"]],
+    [:tar_gz,  Releasy::Packagers::TarGzip,  %[7za a -so -mmt -bd -ttar "f.tar" "f" | 7za a -si -bd -tgzip -mx9 "f.tar.gz"]],
+    [:zip,     Releasy::Packagers::Zip,      %[7za a -mmt -bd -tzip -mx9 "f.zip" "f"]],
 ].each do |type, packager, command|
   extension = "." + type.to_s.tr("_", ".")
 
@@ -22,7 +22,12 @@ require File.expand_path("../teststrap", File.dirname(__FILE__))
 
     asserts(:type).equals type
     asserts(:extension).equals extension
-    asserts(:command, "f").equals command
+    context "" do
+      hookup do
+        mock(Kernel, :`).with("#{Gem.win_platform? ? "where" : "which"} 7za").returns(true) unless type == :dmg
+      end
+      asserts(:command, "f").equals command
+    end
     asserts(:package, "f").equals "f#{extension}"
 
     context "setting extension with a . changes package path" do