+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.2.1
 ------
 
   * Added new deployers: :rsync and :local
+  * Removed 7z installation dependency on Windows (included 7za.exe in gem).
 
 v0.2.0
 ------